requirements.txt
beautifulsoup4==4.5.3
bs4==0.0.1
certifi==2017.7.27.1
chardet==3.0.4
Django==1.10.5
django-activatable-model==0.7.3
django-admin-interface==0.3.1
django-admin-list-filter-dropdown==1.0.1
django-analytical==2.2.2
django-bootstrap3==8.1.0
django-ckeditor==5.3.0
django-colorfield==0.1.12
django-constance==2.0.0
django-extensions==1.7.5
django-flat-responsive==1.3.0
django-flat-theme==1.1.4
django-ipware==1.1.6
django-js-asset==0.1.1
django-kronos==1.0
django-manager-utils==0.12.0
django-picklefield==0.3.2
django-query-builder==0.14.0
django-recaptcha==1.3.1
django-smart-selects==1.3.4
fleming==0.4.4
idna==2.5
olefile==0.43
Pillow==4.0.0
psycopg2==2.7.3.1
python-crontab==2.2.2
python-dateutil==2.6.0
pytz==2017.2
requests==2.12.4
six==1.10.0
urllib3==1.22
